3/2/EYAZIS/3/requirements.txt
spacy
nltk
lemminflect
pdfplumber
python-docx
striprtf
SQLAlchemy
pandas
tqdm
rich
wordfreq
fastapi
uvicorn
streamlit
requests
python-multipart
